+++ packages/opentelemetry-core/src/common/types.ts
@@ -40,11 +40,23 @@ export interface ShimWrapped extends Function {
 }
 
 /**
- * An instrumentation library consists of the name and and optional version
+ * An instrumentation library consists of the name and optional version
+ * used to obtain a tracer or meter from a provider. This metadata is made
+ * available on ReadableSpan and MetricRecord for use by the export pipeline.
+ * @deprecated Use {@link InstrumentationScope} instead.
+ */
+export interface InstrumentationLibrary {
+  readonly name: string;
+  readonly version?: string;
+  readonly schemaUrl?: string;
+}
+
+/**
+ * An instrumentation scope consists of the name and optional version
  * used to obtain a tracer or meter from a provider. This metadata is made
  * available on ReadableSpan and MetricRecord for use by the export pipeline.
  */
-export interface InstrumentationLibrary {
+export interface InstrumentationScope {
   readonly name: string;
   readonly version?: string;
   readonly schemaUrl?: string;